# fast-admin/src/main/resources/public/swagger/index.yaml
swagger: '2.0'
info:
  description: 是一套轻量级权限管理系统，极低门槛，拿来即用。设计之初，就非常注重安全性，为企业系统保驾护航，让一切都变得如此简单。
  version: 3.0.0
  title: universal-framework
basePath: /universal-framework
schemes:
  - http

#定义接口数据
paths:
...
  /sys/role/list:
    get:
      tags:
        - 角色管理
      summary: 角色列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: roleName
          description: 角色名
          in: query
          type: string
      responses:
        '200':
          description: 返回管理员列表
          schema:
            $ref: '#/definitions/SysRoleEntity'
...
  /sys/role/info/{roleId}:
    get:
      tags:
        - 角色管理
      summary: 获取角色信息
      produces:
        - application/json
      parameters:
        - name: roleId
          description: 角色ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回角色信息
          schema:
            $ref: '#/definitions/SysRoleEntity'
...
definitions:
...
  SysRoleEntity:
    type: object
    properties:
      roleId:
        description: 角色ID
        type: integer
        format: int64
      roleName:
        description: 角色名称
        type: string
      deptId:
        description: 部门ID
        type: integer
        format: int64
      deptName:
        description: 部门名称
        type: string
      remark:
        description: 备注
        type: string
      menuIdList:
        description: 菜单ID列表
        type: array
        items:
          type: integer
          format: int64
      createUserId:
        description: 创建者ID
        type: integer
        format: int64
      createTime:
        description: 创建时间
        type: string
        format: date-time
...